BellmanFordAlgorithm.java
The bellman-Ford algorithm helps to find the shortest distance from the source node to all other nodes. It is helpful when Dijikstra's fails.
If the graph contains negative edges.
If the graph has a negative cycle (In this case Dijkstra's algorithm fails to minimize the distance, keeps on running, and goes into an infinite loop. As a result it gives TLE error).